requirements.txt
black==24.2.0
certifi==2024.2.2
click==8.1.7
dnspython==2.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
pymongo==4.6.1
python-dotenv==1.0.1
tomli==2.0.1
typing_extensions==4.9.0
